fix: remove distro_arch variable (#20)
The `distro_arch` variable defined in `vars/main.yml` was used by the
`mongo_toolchain_url` variable defined in `defaults/main.yml`.

The `mongo_toolchain_url` variable contained a statement:

`{% if ansible_architecture != 'x86_64' %}{{ distro_arch }}-{% endif %}`

Which reads, if the architecture is not x86_64, then use the value of
`distro_arch`. However, the `distro_arch` variable is created in memory
at runtime.

So if the architecture is x86_64, then `distro_arch` attempts to lookup
the key (`x86_64`) in the `distro_arch_dict` dictionary, which does not
exist.

Instead of adding an `x86_64` key to the `distro_arch_dict` dictionary
with the value of `""`, I removed the `distro_arch` variable entirely
and substituted it in the `mongo_toolchain_url` variable.

Considering the statement explicitly ignores x86_64 architectures, and
was only referenced by the `mongo_toolchain_url` variable, it makes more
sense to remove the variable entirely.